database-filler/requirements.txt
pandas==1.3.3
psycopg2-binary==2.9.1
sqlalchemy==1.4.26
